requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
langchain==0.3.3
pandas==2.2.3
pymongo==4.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.39.0
Werkzeug==3.0.4
streamlit   
langchain-community    
sentence-transformers     
faiss-cpu